# payments/docs/swagger.yaml
basePath: /payment/
definitions:
  CreateAccountRequest:
    properties:
      balance:
        example: 1000
        type: number
      full_name:
        example: John Doe
        type: string
    type: object
  UpdateAccountRequest:
    properties:
      amount:
        example: 100
        type: number
    type: object
  model.Account:
    properties:
      balance:
        type: number
      full_name:
        type: string
      user_id:
        type: string
    type: object
host: localhost:8081
info:
  contact: {}
  title: Payments
  version: "1.0"
paths:
  /account/create/{user_id}:
    post:
      description: создает аккаунта
      parameters:
      - description: ID пользователя
        in: path
        name: user_id
        required: true
        type: string
      - description: Начальный счет аккаунта
        in: body
        name: balance
        required: true
        schema:
          $ref: '#/definitions/CreateAccountRequest'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      tags:
      - Accounts Manage
  /account/get:
    get:
      description: Возвращает все счета
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Account'
            type: array
      tags:
      - Accounts Info
  /account/get/{user_id}:
    get:
      description: Получить счет по ID
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Информация о счете
          schema:
            $ref: '#/definitions/model.Account'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      tags:
      - Accounts Info
  /account/update/{user_id}:
    patch:
      description: изменить баланс счета
      parameters:
      - description: ID пользователя
        in: path
        name: user_id
        required: true
        type: string
      - description: Изменение счета
        in: body
        name: update_request
        required: true
        schema:
          $ref: '#/definitions/UpdateAccountRequest'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      tags:
      - Accounts Manage
swagger: "2.0"